+++ src/GraphicsAPI/Metal/Display.h
@@ -15,7 +15,7 @@ public:
     static void draw_cycle_callback();
     static void key_up_callback(int a, int b);
     static void key_down_callback(int a, int b);
-    static void mouse_move_callback(float a, float b);
+    static void mouse_move_callback(int a, int b);
     static void mouse_down_callback(int a, int b);
 };
